feat: support reading assets from client 11 or higher (#70)
feat: migrate Remeres to 12.x assets system and modernized item handling

• Migrated old Remeres .dat/.spr system to new 12.x client /assets/ appearance & sprite format
• Removed OTB dependency and outdated client version structures
• Integrated Protobuf-based sprite reader (credit: @nekiro)
• Major refactor on graphics.cpp structure and brush icon handling
• Updated brushes and tags system
• Fixed top order drawing bugs and missing brushes icons
• Fixed outfit color paint and correct offset handling for non-32x32 sprites
• Re-enabled map saving and loading
• Improved creature sprite rendering and brush previews
• Added new item properties window with toggle for legacy/new layout

+++ source/monster.h
@@ -65,10 +65,10 @@ public:
 		weight = newWeight;
 	}
 
-	int getSpawnMonsterTime() const noexcept {
+	uint16_t getSpawnMonsterTime() const noexcept {
 		return spawntime;
 	}
-	void setSpawnMonsterTime(int time) noexcept {
+	void setSpawnMonsterTime(uint16_t time) noexcept {
 		spawntime = time;
 	}
 
@@ -87,7 +87,7 @@ protected:
 	std::string type_name;
 	Direction direction;
 	uint8_t weight;
-	int spawntime;
+	uint16_t spawntime;
 	bool saved;
 	bool selected;
 };
